+++ projection/src/main/scala/org/apache/pekko/projection/r2dbc/javadsl/R2dbcSession.scala
@@ -36,11 +36,11 @@ final class R2dbcSession(connection: Connection)(implicit ec: ExecutionContext, 
   def createStatement(sql: String): Statement =
     connection.createStatement(sql)
 
-  def updateOne(statement: Statement): CompletionStage[Integer] =
-    R2dbcExecutor.updateOneInTx(statement).map(Integer.valueOf)(ExecutionContexts.parasitic).asJava
+  def updateOne(statement: Statement): CompletionStage[java.lang.Long] =
+    R2dbcExecutor.updateOneInTx(statement).map(java.lang.Long.valueOf)(ExecutionContexts.parasitic).asJava
 
-  def update(statements: java.util.List[Statement]): CompletionStage[java.util.List[Integer]] =
-    R2dbcExecutor.updateInTx(statements.asScala.toVector).map(results => results.map(Integer.valueOf).asJava).asJava
+  def update(statements: java.util.List[Statement]): CompletionStage[java.util.List[java.lang.Long]] =
+    R2dbcExecutor.updateInTx(statements.asScala.toVector).map(results => results.map(java.lang.Long.valueOf).asJava).asJava
 
   def selectOne[A](statement: Statement)(mapRow: Row => A): CompletionStage[Optional[A]] =
     R2dbcExecutor.selectOneInTx(statement, mapRow).map(_.toJava)(ExecutionContexts.parasitic).asJava
